Exception handling

+++ src/main/java/com/luisbicho/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.luisbicho.dscommerce.services;
 import com.luisbicho.dscommerce.dto.ProductDTO;
 import com.luisbicho.dscommerce.entities.Product;
 import com.luisbicho.dscommerce.repositories.ProductRepository;
+import com.luisbicho.dscommerce.services.exceptions.DatabaseException;
+import com.luisbicho.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -20,7 +25,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
         Optional<Product> product = productRepository.findById(id);
-        return new ProductDTO(product.get());
+        return new ProductDTO(product.orElseThrow(() -> new ResourceNotFoundException("ID Not Found")));
     }
 
     @Transactional(readOnly = true)
@@ -39,16 +44,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
